Applying turn_bq extension logic to other applicable submodules.
