+++ rdw_ingestion_tools/api/rapidpro/requests/flow_starts.py
@@ -1,8 +1,9 @@
 from attrs import define
 from httpx import Client
-from pandas import DataFrame, concat, json_normalize
+from pandas import DataFrame
 
-from .. import get_paginated
+from ..extensions.dataframe import concatenate
+from ..extensions.httpx import get_paginated
 
 
 @define
@@ -27,14 +28,6 @@ class FlowStarts:
 
         flowstarts_generator = get_paginated(self.client, url, **kwargs)
 
-        flowstarts_list: list[DataFrame] = [
-            json_normalize(response, sep="_")
-            for response in flowstarts_generator
-        ]
-
-        try:
-            flowstarts = concat(flowstarts_list)
-        except ValueError:
-            flowstarts = DataFrame()
+        flowstarts = concatenate(flowstarts_generator)
 
         return flowstarts
