+++ rdw_ingestion_tools/api/rapidpro/requests/runs.py
@@ -1,8 +1,9 @@
 from attrs import define
 from httpx import Client
-from pandas import DataFrame, concat, json_normalize
+from pandas import DataFrame
 
-from .. import get_paginated
+from ..extensions.dataframe import concatenate
+from ..extensions.httpx import get_paginated
 
 
 @define
@@ -27,13 +28,6 @@ class Runs:
 
         runs_generator = get_paginated(self.client, url, **kwargs)
 
-        runs_list: list[DataFrame] = [
-            json_normalize(response, sep="_") for response in runs_generator
-        ]
-
-        try:
-            runs = concat(runs_list)
-        except ValueError:
-            runs = DataFrame()
+        runs = concatenate(runs_generator)
 
         return runs
